--- src/corenodes/features/__init__.py
@@ -1,2 +1,0 @@
-from src.corenodes.features.resize import ResizeModule
-from src.corenodes.features.rotate import RotateModule
